Angular: Isolate input transform analysis
